Refresh refactored and now verifies that the token is valid - more tests needed
