+++ test/refresh.tests.js
@@ -92,32 +92,37 @@ describe('Refresh Token Testing', function() {
   });
 
   it('Should be able to refresh the token', function (done) {
-      var refreshed = jwt.refresh(jwt.decode(token, {complete: true}), 3600, secret);
+
+      var refreshed = jwt.refresh(token, 3600, secret);
       assert.ok(refreshed);
       done();
   });
 
   it('Should be able to refresh the token (async)', function (done) {
-      var refreshed = jwt.refresh(jwt.decode(token, {complete: true}), 3600, secret, function(err, refreshedToken) {
+
+      jwt.refresh(token, 3600, secret, null, function(err, refreshedToken) {
+
           assert.ok(refreshedToken);
+          done();
       });
-      done();
   });
 
   var originalDecoded = jwt.decode(token, {complete: true});
-  var refreshed = jwt.refresh(originalDecoded, 3600, secret);
+  var refreshed = jwt.refresh(token, 3600, secret);
   var refreshDecoded = jwt.decode(refreshed, {complete: true});
   var refreshAsync;
   var refreshAsyncDecoded;
-  jwt.refresh(jwt.decode(token, {complete: true}), 3600, secret, function(err, refreshedToken) {
+  jwt.refresh(token, 3600, secret, null, function(err, refreshedToken) {
+
       refreshAsync = refreshedToken;
-      refreshAsyncDecoded = jwt.decode(refreshed, {complete: true});
+      refreshAsyncDecoded = jwt.decode(refreshedToken, {complete: true});
   });
 
   it('Sub-test to ensure that the compare method works', function (done) {
+
       var originalMatch = equal(originalDecoded, originalDecoded);
       var refreshMatch = equal(refreshDecoded, refreshDecoded);
-      var asyncRefreshMatch = equal(originalDecoded, refreshAsyncDecoded);
+      var asyncRefreshMatch = equal(refreshAsyncDecoded, refreshAsyncDecoded);
 
       assert.equal(originalMatch, refreshMatch);
       assert.equal(originalMatch, asyncRefreshMatch);
@@ -125,6 +130,7 @@ describe('Refresh Token Testing', function() {
   });
 
   it('Decoded version of a refreshed token should be the same, except for timing data', function (done) {
+
       var comparison = equal(originalDecoded, refreshDecoded);
       var asyncComparison = equal(originalDecoded, refreshAsyncDecoded);
 
@@ -134,43 +140,13 @@ describe('Refresh Token Testing', function() {
   });
 
   it('Refreshed token should have a later expiery time then the original', function (done) {
+
       var originalExpiry = originalDecoded.payload.exp;
       var refreshedExpiry = refreshDecoded.payload.exp;
       var refreshedAsyncExpiry = refreshAsyncDecoded.payload.exp;
 
       assert.isTrue((refreshedExpiry > originalExpiry), 'Refreshed expiry time is above original time');
       assert.isTrue((refreshedAsyncExpiry > originalExpiry), 'Refreshed expiry time is above original time (async)');
-      done();
-  });
-
-  it('Refreshing a token that\'s is not from an original decoded token should still work - basically creating a brand new token', function (done) {
-      var notReallyAToken = {
-          key: 'value',
-          foo: 'bar',
-          not: 'a token'
-      }
-      var notReallyATokenRefresh = jwt.refresh(notReallyAToken, 3600, secret);
-
-      assert.ok(notReallyATokenRefresh);
-      done();
-  });
-
-  it('Should fail when not providing a time value for the expiresIn value', function (done) {
-      var notReallyAToken = {
-          key: 'value',
-          foo: 'bar',
-          not: 'a token'
-      }
-
-      var failRefresh;
-      try {
-          var failRefresh = jwt.refresh(notReallyAToken, null, secret);
-      } catch (err) {
-          assert.equal(err.name, 'ValidationError');
-          assert.equal(err.message, 'child "expiresIn" fails because ["expiresIn" must be a number, "expiresIn" must be a string]');
-      }
-
-      assert.notOk(failRefresh);
       done();
   });
 });
